refactor: 읽기 편하게 리팩토링
