--- src/main/java/com/rudkids/rudkids/domain/cart/application/CartItemStore.java
@@ -1,7 +1,0 @@
-package com.rudkids.rudkids.domain.cart.application;
-
-import com.rudkids.rudkids.domain.cart.domain.CartItem;
-
-public interface CartItemStore {
-    CartItem store(CartItem cartItem);
-}
