--- src/main/java/com/rudkids/rudkids/domain/cart/application/CartStore.java
@@ -1,7 +1,0 @@
-package com.rudkids.rudkids.domain.cart.application;
-
-import com.rudkids.rudkids.domain.cart.domain.Cart;
-
-public interface CartStore {
-    Cart store(Cart cart);
-}
