+++ src/main/java/com/rudkids/rudkids/infrastructure/cart/CartItemReaderImpl.java
@@ -8,15 +8,20 @@ import com.rudkids.rudkids.domain.item.domain.Item;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
-import java.util.Optional;
-
 @Component
 @RequiredArgsConstructor
 public class CartItemReaderImpl implements CartItemReader {
     private final CartItemRepository cartItemRepository;
 
     @Override
-    public Optional<CartItem> findCartItem(Cart cart, Item item) {
-        return cartItemRepository.findByCartAndItem(cart, item);
+    public CartItem getCartItem(Cart cart, Item item) {
+        return cartItemRepository.findByCartAndItem(cart, item)
+                .orElseGet(() -> createCartItem(cart, item));
+    }
+
+    private CartItem createCartItem(Cart cart, Item item) {
+        var cartItem = CartItem.create(cart, item);
+        cart.addCartItem(cartItem);
+        return cartItemRepository.save(cartItem);
     }
 }
